Сделано: 1. Создал класс CheckFieldsInClass, использующий Bean Validation для проверки полей любых объектов 2. Создал тест CheckFieldsInClassTest 3. Добавил везде класс CheckFieldsInClass, где конструкторы или методы принимают на вход классы 4. заменил все поля про деньги с float на long (решил не использовать BigDecimal) - в идеале надо бы ввести поля с копейками, а рубли внутри методов умножать на 100, но мне пока влом, т.к. в целях обучения не несет смысловой нагрузки, а времени уйдет много для правки кода и тестов
Необходимо сделать:
1. реализовать логирование при перехвате исключений, вместо sout

+++ src/main/java/ru/sbrf/payment/client/AccountDebit.java
@@ -7,7 +7,7 @@ import ru.sbrf.payment.common.Currency;
 
 public class AccountDebit extends Account {
 
-    public AccountDebit(String accountNumber, Currency currency, float balance) {
+    public AccountDebit(String accountNumber, Currency currency, long balance) {
         super(accountNumber, currency, balance);
     }
